pyproject.toml
[tool.poetry]
name = "interview-tasks"
version = "0.1.0"
description = ""
authors = ["Vladimir Molchanov <r1ken@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
fastapi = "^0.115.6"
alembic = "^1.14.0"
pydantic-settings = "^2.7.0"
aiosqlite = "^0.20.0"
greenlet = "^3.1.1"
jinja2 = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
